JessamineBenchmark

Documentation for JessamineBenchmark.

JessamineBenchmark.FriedmanFunctions

JessamineBenchmark.SinFunction

JessamineBenchmark.FriedmanFunctions.friedman1

JessamineBenchmark.FriedmanFunctions.friedman2

JessamineBenchmark.FriedmanFunctions.friedman3

JessamineBenchmark.FriedmanFunctions.make_rand_friedman1

JessamineBenchmark.FriedmanFunctions.make_rand_friedman2

JessamineBenchmark.FriedmanFunctions.make_rand_friedman3

JessamineBenchmark.SinFunction.make_rand_sin

JessamineBenchmark.SinFunction — Module

Generate data from sin

To run this CLI app manually:

julia --project=@. -m JessamineBenchmark.SinFunction

JessamineBenchmark.SinFunction.make_rand_sin — Method

 make_rand_sin(kwargs...)

Make random samples of the input $x$ as columns and apply sin to the result. Make additional random samples as columns to use as distractors.

Keywords

rng=Random.default_rng(): Random number generator to use.

num_samples=100: How many points (rows) to create.

input_dist=Uniform(-pi, pi): Distribution of inputs.

noise_dist=Dirac(0): Distribution of noise; default means no noise.

Returns

A DataFrame with the following columns:

y: result of applying sin to input x plus noise.

x: input

JessamineBenchmark.FriedmanFunctions — Module

Generate data from Friedman functions, which are commonly used for machine learning benchmarks.

In Friedman's original article, friedman1 is (61), friedman2 is (63a), and friedman3 is (63b).

Breiman, Leo. Bagging Predictors. Machine Learning, v24 no2 p123–140, 1996-08-01. doi:10.1007/BF00058655

Friedman, Jerome H. Multivariate adaptive regression splines. The Annals of Statistics v19 no1 p1–67, 1991. JStor

This module follows the Scikit-learn implementation

To run this CLI app manually:

julia --project=@. -m JessamineBenchmark.FriedmanFunctions

JessamineBenchmark.FriedmanFunctions.friedman1 — Method

friedman1(x1, x2, x3, x4, x5)

Return $10 \sin(π x_1 x_2) + 20 (x_3 - 1/2)^2 + 10 x_4 + 5 x_5$.

JessamineBenchmark.FriedmanFunctions.friedman2 — Method

friedman2(x1, x2, x3, x4)

Return $\sqrt{x_1^2 + \left(x_2 x_3 - \frac{1}{x_2 x4}\right)}$.

JessamineBenchmark.FriedmanFunctions.friedman3 — Method

friedman3(x1, x2, x3, x4)

Return $\atan \left(\frac{x_2 x_3 - 1/(x_2 x_4)}{x_1}\right)$.

JessamineBenchmark.FriedmanFunctions.make_rand_friedman1 — Method

 make_rand_friedman1(kwargs...)

Make random samples of the inputs $x_1, \dots, x_5$ as columns and apply friedman1 to the result. Make additional random samples as columns to use as distractors.

Keywords

rng=Random.default_rng(): Random number generator to use.

num_samples=100: How many points (rows) to create.

num_distractors=0: How many distractor columns to create.

input_dist=Uniform(): Distribution of inputs and distractors.

noise_dist=Dirac(0): Distribution of noise; default means no noise.

Returns

A DataFrame with the following columns:

y: result of applying friedman1 to inputs x1, ..., x5, plus noise.

x1, ..., x5: the five inputs

x6, ... : distractors

JessamineBenchmark.FriedmanFunctions.make_rand_friedman2 — Method

make_rand_friedman2(kwargs...)

Make random samples of the inputs $x_1, \dots, x_5$ as columns and apply friedman2 to the result. Make additional random samples as columns to use as distractors.

Keywords

rng=Random.default_rng(): Random number generator to use.

num_samples=100: How many points (rows) to create.

x1_dist=Uniform(0, 100): Distribution for x1.

x2_dist=Uniform(2*π*40, 2*π*280): Distribution for x2.

x3_dist=Uniform(0, 1): Distribution for x3

x4_dist=Uniform(1, 11): Distribution for x4.

noise_dist=Dirac(0): Distribution of noise; default means no noise.

Returns

A DataFrame with the following columns:

y: result of applying friedman2 to inputs x1, ..., x4, plus noise.

x1, ..., x4: the four inputs

JessamineBenchmark.FriedmanFunctions.make_rand_friedman3 — Method

make_rand_friedman3(kwargs...)

Make random samples of the inputs $x_1, \dots, x_5$ as columns and apply friedman3 to the result. Make additional random samples as columns to use as distractors.

Keywords

rng=Random.default_rng(): Random number generator to use.

num_samples=100: How many points (rows) to create.

x1_dist=Uniform(0, 100): Distribution for x1.

x2_dist=Uniform(2*π*40, 2*π*280): Distribution for x2.

x3_dist=Uniform(0, 1): Distribution for x3

x4_dist=Uniform(1, 11): Distribution for x4.

noise_dist=Dirac(0): Distribution of noise; default means no noise.

Returns

A DataFrame with the following columns:

y: result of applying friedman3 to inputs x1, ..., x4, plus noise.

x1, ..., x4: the four inputs